+++ projectroles/email.py
@@ -449,6 +449,23 @@ def get_user_addr(user):
     return ret
 
 
+def get_project_modify_recipients(project, sender):
+    """
+    Return owner and delegate recipients for project modify emails. Excludes
+    sender user, inactive users and users with notify_email_project set False.
+
+    :param project: Project object
+    :param sender: User object
+    """
+    return [
+        a.user
+        for a in project.get_roles(max_rank=ROLE_RANKING[PROJECT_ROLE_DELEGATE])
+        if a.user != sender
+        and a.user.is_active
+        and app_settings.get(APP_NAME, 'notify_email_project', user=a.user)
+    ]
+
+
 def send_mail(
     subject,
     message,
@@ -649,106 +666,110 @@ def send_invite_expiry_mail(invite, request, user_name):
 
 def send_project_create_mail(project, request):
     """
-    Send email about project creation to the owner of the parent category, if
-    they are a different user than the project creator.
+    Send email about project creation to owners and delegates of the parent
+    category, excluding the project creator. Also excludes users who have set
+    notify_email_role to False.
 
     :param project: Project object for the newly created project
     :param request: HttpRequest object
     :return: Amount of sent email (int)
     """
     parent = project.parent
-    parent_owner = project.parent.get_owner() if project.parent else None
-    project_owner = project.get_owner()
-    if (
-        not parent
-        or not parent_owner
-        or not parent_owner.user.is_active
-        or parent_owner.user == request.user
-    ):
+    if not parent:
         return 0
+    recipients = get_project_modify_recipients(project, request.user)
+    if not recipients:
+        return 0
+    owner = project.get_owner().user
 
     subject = SUBJECT_PREFIX + SUBJECT_PROJECT_CREATE.format(
         project_type=get_display_name(project.type, title=True),
         project=project.title,
         user=get_email_user(request.user),
     )
-    message = get_email_header(
-        MESSAGE_HEADER.format(
-            recipient=parent_owner.user.get_full_name(), site_title=SITE_TITLE
-        )
-    )
-    message += MESSAGE_PROJECT_CREATE_BODY.format(
-        user=get_email_user(request.user),
-        project_type=get_display_name(project.type),
-        category=parent.title,
-        project=project.title,
-        owner=get_email_user(project_owner.user),
-        project_url=request.build_absolute_uri(
-            reverse(
-                'projectroles:detail', kwargs={'project': project.sodar_uuid}
+    mail_count = 0
+    for r in recipients:
+        message = get_email_header(
+            MESSAGE_HEADER.format(
+                recipient=r.get_full_name(), site_title=SITE_TITLE
             )
-        ),
-    )
-    message += get_email_footer(request)
-    return send_mail(
-        subject,
-        message,
-        get_user_addr(parent_owner.user),
-        request,
-        reply_to=get_user_addr(request.user),
-    )
+        )
+        message += MESSAGE_PROJECT_CREATE_BODY.format(
+            user=get_email_user(request.user),
+            project_type=get_display_name(project.type),
+            category=parent.title,
+            project=project.title,
+            owner=get_email_user(owner),
+            project_url=request.build_absolute_uri(
+                reverse(
+                    'projectroles:detail',
+                    kwargs={'project': project.sodar_uuid},
+                )
+            ),
+        )
+        message += get_email_footer(request)
+        mail_count += send_mail(
+            subject,
+            message,
+            get_user_addr(r),
+            request,
+            reply_to=get_user_addr(request.user),
+        )
+    return mail_count
 
 
 def send_project_move_mail(project, request):
     """
-    Send email about project being moved to the owner of the parent category, if
-    they are a different user than the project creator.
+    Send email about project being moved to the owners and delegates of the
+    parent category, excluding the project creator. Also excludes users who have
+    set notify_email_role to False.
 
     :param project: Project object for the newly created project
     :param request: HttpRequest object
     :return: Amount of sent email (int)
     """
     parent = project.parent
-    parent_owner = project.parent.get_owner() if project.parent else None
-    project_owner = project.get_owner()
-    if (
-        not parent
-        or not parent_owner
-        or not parent_owner.user.is_active
-        or parent_owner.user == request.user
-    ):
+    if not parent:
         return 0
+    recipients = get_project_modify_recipients(project, request.user)
+    if not recipients:
+        return 0
+    owner = project.get_owner().user
 
     subject = SUBJECT_PREFIX + SUBJECT_PROJECT_MOVE.format(
         project_type=get_display_name(project.type, title=True),
         project=project.title,
         user=get_email_user(request.user),
     )
-    message = get_email_header(
-        MESSAGE_HEADER.format(
-            recipient=parent_owner.user.get_full_name(), site_title=SITE_TITLE
-        )
-    )
-    message += MESSAGE_PROJECT_MOVE_BODY.format(
-        user=get_email_user(request.user),
-        project_type=get_display_name(project.type),
-        category=parent.title,
-        project=project.title,
-        owner=get_email_user(project_owner.user),
-        project_url=request.build_absolute_uri(
-            reverse(
-                'projectroles:detail', kwargs={'project': project.sodar_uuid}
+    mail_count = 0
+    for r in recipients:
+        message = get_email_header(
+            MESSAGE_HEADER.format(
+                recipient=r.get_full_name(), site_title=SITE_TITLE
             )
-        ),
-    )
-    message += get_email_footer(request)
-    return send_mail(
-        subject,
-        message,
-        get_user_addr(parent_owner.user),
-        request,
-        reply_to=get_user_addr(request.user),
-    )
+        )
+        message += MESSAGE_PROJECT_MOVE_BODY.format(
+            user=get_email_user(request.user),
+            project_type=get_display_name(project.type),
+            category=parent.title,
+            project=project.title,
+            owner=get_email_user(owner),
+            project_url=request.build_absolute_uri(
+                reverse(
+                    'projectroles:detail',
+                    kwargs={'project': project.sodar_uuid},
+                )
+            ),
+        )
+        message += get_email_footer(request)
+        mail_count += send_mail(
+            subject,
+            message,
+            get_user_addr(r),
+            request,
+            reply_to=get_user_addr(request.user),
+        )
+    return mail_count
 
 
 def send_project_archive_mail(project, action, request):
@@ -760,16 +781,14 @@ def send_project_archive_mail(project, action, request):
     :param request: HttpRequest object
     :return: Amount of sent email (int)
     """
-    user = request.user
-    project_users = [
+    recipients = [
         a.user
         for a in project.get_roles()
-        if a.user != user
+        if a.user != request.user
         and a.user.is_active
         and app_settings.get(APP_NAME, 'notify_email_project', user=a.user)
     ]
-    project_users = list(set(project_users))  # Remove possible dupes (see #710)
-    if not project_users:
+    if not recipients:
         return 0
     mail_count = 0
 
@@ -782,13 +801,13 @@ def send_project_archive_mail(project, action, request):
     subject = SUBJECT_PREFIX + subject.format(
         project_label_title=get_display_name(project.type, title=True),
         project=project.title,
-        user=user.get_full_name(),
+        user=request.user.get_full_name(),
     )
     body_final = body.format(
         project_label_title=get_display_name(project.type, title=True),
         project_label=get_display_name(project.type),
         project=project.title,
-        user=user.get_full_name(),
+        user=request.user.get_full_name(),
         project_url=request.build_absolute_uri(
             reverse(
                 'projectroles:detail', kwargs={'project': project.sodar_uuid}
@@ -796,17 +815,17 @@ def send_project_archive_mail(project, action, request):
         ),
     )
 
-    for recipient in project_users:
+    for r in recipients:
         message = get_email_header(
             MESSAGE_HEADER.format(
-                recipient=recipient.get_full_name(), site_title=SITE_TITLE
+                recipient=r.get_full_name(), site_title=SITE_TITLE
             )
         )
         message += body_final
         if not settings.PROJECTROLES_EMAIL_SENDER_REPLY:
             message += NO_REPLY_NOTE
         message += get_email_footer(request)
-        mail_count += send_mail(subject, message, get_user_addr(user), request)
+        mail_count += send_mail(subject, message, get_user_addr(r), request)
     return mail_count
 
 
@@ -819,15 +838,15 @@ def send_project_delete_mail(project, request):
     :return: Amount of sent email (int)
     """
     user = request.user
-    project_users = [
+    recipients = [
         a.user
         for a in project.get_roles()
         if a.user != user
         and a.user.is_active
         and app_settings.get(APP_NAME, 'notify_email_project', user=a.user)
     ]
-    project_users = list(set(project_users))
-    if not project_users:
+    recipients = list(set(recipients))
+    if not recipients:
         return 0
 
     mail_count = 0
@@ -844,17 +863,17 @@ def send_project_delete_mail(project, request):
         user=user.get_full_name(),
     )
 
-    for recipient in project_users:
+    for r in recipients:
         message = get_email_header(
             MESSAGE_HEADER.format(
-                recipient=recipient.get_full_name(), site_title=SITE_TITLE
+                recipient=r.get_full_name(), site_title=SITE_TITLE
             )
         )
         message += body_final
         if not settings.PROJECTROLES_EMAIL_SENDER_REPLY:
             message += NO_REPLY_NOTE
         message += get_email_footer(request)
-        mail_count += send_mail(subject, message, get_user_addr(user), request)
+        mail_count += send_mail(subject, message, get_user_addr(r), request)
     return mail_count
 
 
@@ -867,7 +886,6 @@ def send_generic_mail(
     cc=None,
     bcc=None,
     settings_link=True,
-    support_email=False,
 ):
     """
     Send a generic mail with standard header and footer and no-reply
@@ -881,7 +899,6 @@ def send_generic_mail(
     :param cc: List of emails for "cc" field (optional)
     :param bcc: List of emails for "bcc" field (optional)
     :param settings_link: Include link to user settings if True (optional)
-    :param support_email: Include support contact info if True (optional)
     :return: Amount of mail sent (int)
     """
     subject = SUBJECT_PREFIX + subject_body